requirements.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.5
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
flower==0.9.4
future==0.18.2
humanize==0.5.1
idna==2.9
kombu==4.6.8
psycopg2-binary==2.8.5
PyJWT==1.7.1
python-dotenv==0.13.0
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
tornado==6.0.4
urllib3==1.25.8
vine==1.3.0
